wip: merge signer code, rename test package, handle errors
